// src/Anfx.Pasivos.Frontend/src/api/models/contratoPasivoDto.ts
/**
 * Anfx.Pasivos.ApiService
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PagoIrregularDto } from './pagoIrregularDto';


export interface ContratoPasivoDto { 
    fondeador?: string | null;
    idLineaCredito?: number;
    maxCapitalDisponible?: number;
    idFondeador?: number;
    idTipoCredito?: number | null;
    contrato?: string | null;
    idEstatusContrato?: number | null;
    capitalFinanciado?: number | null;
    plazo?: number | null;
    idMoneda?: number;
    idPeriodicidad?: number;
    tipoTasa?: boolean | null;
    idTasa?: number;
    tasaIva?: number | null;
    tasaBase?: number | null;
    tasa?: number | null;
    puntosMas?: number | null;
    puntosPor?: number | null;
    tipoTasaMora?: boolean | null;
    idTasaMora?: number;
    tasaMora?: number | null;
    tasaBaseMora?: number | null;
    puntosMasMora?: number | null;
    factorMora?: number | null;
    fechaFirmaContrato?: string | null;
    fecInicioContrato?: string | null;
    fecPrimeraRenta?: string | null;
    fecActivacion?: string | null;
    fecFinContrato?: string | null;
    idTipoTablaAmortiza?: number;
    idTipoCapitalizacion?: number | null;
    idTipoPagoCapital?: number | null;
    idPeriodicidad_TTA?: number | null;
    noPagosIrregulares?: number | null;
    tipoCredito?: string | null;
    pagos?: Array<PagoIrregularDto> | null;
}