// src/Anfx.Pasivos.Frontend/src/api/models/tipoTablaAmortizaListItemDto.ts
/**
 * Anfx.Pasivos.ApiService
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface TipoTablaAmortizaListItemDto { 
    id?: number;
    tipoTablaAmortiza?: string | null;
    activo?: boolean;
}